half finish server
