--- src/server/index.js
@@ -1,3 +1,0 @@
-import createApp from './createApp'
-
-export default createApp
