--- src/server/viewEngine.js
@@ -1,5 +1,0 @@
-/**
- * default view engine for server
- */
-
-export let render = html => html.toString()
