+++ src/server/viewEngine.ts
@@ -1,0 +1,9 @@
+/**
+ * default view engine for server
+ */
+import { ViewEngine } from '../share/constant'
+const render = html => html.toString()
+
+export default {
+  render
+}
